+++ fuel-vm/src/memory_client.rs
@@ -3,22 +3,24 @@
 use crate::{
     backtrace::Backtrace,
     checked_transaction::Checked,
+    error::InterpreterError,
+    interpreter::{
+        EcalHandler,
+        InterpreterParams,
+        NotSupportedEcal,
+    },
     state::StateTransitionRef,
     storage::MemoryStorage,
     transactor::Transactor,
 };
-
-use crate::interpreter::{
-    EcalHandler,
-    InterpreterParams,
-    NotSupportedEcal,
-};
+use core::convert::Infallible;
 use fuel_tx::{
     Create,
     FeeParameters,
     GasCosts,
     Receipt,
     Script,
+    Upgrade,
 };
 
 #[derive(Debug)]
@@ -79,8 +81,19 @@ impl<Ecal: EcalHandler> MemoryClient<Ecal> {
     }
 
     /// Deploys a `Create` transaction.
-    pub fn deploy(&mut self, tx: Checked<Create>) -> Option<Create> {
-        self.transactor.deploy(tx).ok()
+    pub fn deploy(
+        &mut self,
+        tx: Checked<Create>,
+    ) -> Result<Create, InterpreterError<Infallible>> {
+        self.transactor.deploy(tx)
+    }
+
+    /// Executes `Upgrade` transaction.
+    pub fn upgrade(
+        &mut self,
+        tx: Checked<Upgrade>,
+    ) -> Result<Upgrade, InterpreterError<Infallible>> {
+        self.transactor.upgrade(tx)
     }
 
     /// Execute a transaction.